Remove sched_elect() in serialchar_callback()
Early triggering scheduler results in failure of
unstacking in exception return.

Unexpected exception return increases the probability
of occurrence of faults, such as INVPC, INVSTATE. In
this situation, we can't make sure that it wouldn't
cause faults or invalid interrupt raising due to
unfinished interrupt handling.

For example, if a context switch is done inside the irq,
it will make v7m-head.S `pop {pc}` to a undefined place.

Closed: #12